StageStructureIDV1.1 : Created Headers for Structure
